Move tracking to a bottom sheet
