+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/track/SetTrackReadingDatesDialog.kt
@@ -15,16 +15,19 @@ import uy.kohesive.injekt.api.get
 import java.util.Calendar
 
 class SetTrackReadingDatesDialog<T> : DialogController
-        where T : Controller, T : SetTrackReadingDatesDialog.Listener {
+        where T : Controller {
 
     private val item: TrackItem
 
     private val dateToUpdate: ReadingDate
 
-    constructor(target: T, dateToUpdate: ReadingDate, item: TrackItem) : super(
+    private lateinit var listener: Listener
+
+    constructor(target: T, listener: Listener, dateToUpdate: ReadingDate, item: TrackItem) : super(
         bundleOf(KEY_ITEM_TRACK to item.track)
     ) {
         targetController = target
+        this.listener = listener
         this.item = item
         this.dateToUpdate = dateToUpdate
     }
@@ -38,8 +41,6 @@ class SetTrackReadingDatesDialog<T> : DialogController
     }
 
     override fun onCreateDialog(savedViewState: Bundle?): Dialog {
-        val listener = (targetController as? Listener)
-
         return MaterialDialog(activity!!)
             .title(
                 when (dateToUpdate) {
