+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/track/TrackAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import eu.kanade.tachiyomi.databinding.TrackItemBinding
 
-class TrackAdapter(controller: TrackController) : RecyclerView.Adapter<TrackHolder>() {
+class TrackAdapter(listener: OnClickListener) : RecyclerView.Adapter<TrackHolder>() {
 
     private lateinit var binding: TrackItemBinding
 
@@ -17,7 +17,7 @@ class TrackAdapter(controller: TrackController) : RecyclerView.Adapter<TrackHold
             }
         }
 
-    val rowClickListener: OnClickListener = controller
+    val rowClickListener: OnClickListener = listener
 
     fun getItem(index: Int): TrackItem? {
         return items.getOrNull(index)
